Colors!
Okay, so I got a bit off track here; I'm colorizing the world.

A simple topView function was what I needed to play with the
"rendering" of the world. Once I got that, I could actually visit the
world and figure out what was where; and I built a big arrow pointing
up; which, in combination with some nearby sugarcane, made it
perfectly obvious how the block was structured.