Beginning to converge on an efficient screen drawing implementation.
It took three days for me to realize that it would be more efficient
to generate a placement function than to do the recalculation at every
chunk.

Given that, the placemenet generator `chunkPlacer` will -- given the
screen size and offset -- translate, crop, and place any chunk that is
given to it without doing the screen calculations again the next step
is to generalize this placer so that it can make pure rendering
instructions which can either be printed, or actually rendered.