Testing rendering of single chunks into the right place.
Note that the row subtraction, and the odd behaviour of clearScreen
with regards to the last character are specifically to avoid a crash
which I have observed when using `drawString` to put a character in
the last spot.  I will research whether this is an ncurses,
haskell-ncurses, or greyson-using-ncurses bug later.